refactor(entity): move related classes to the same file
- reason: to avoid confusion -> Profile is not an entity, it is part of an entity
